refactor(app-shell): Update service event handling in AppShell and component usages
